Main/Mouse: mouse handling fix found by the testcase
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@36150 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Main/src-client/MouseImpl.cpp
@@ -484,8 +484,8 @@ HRESULT Mouse::convertDisplayRes(LONG x, LONG y, int32_t *pcX, int32_t *pcY,
         if (FAILED(rc))
             return rc;
 
-        *pcX = displayWidth ? ((x - 1) * 0xFFFF) / displayWidth: 0;
-        *pcY = displayHeight ? ((y - 1) * 0xFFFF) / displayHeight: 0;
+        *pcX = displayWidth ? ((x - 1) * 0xFFFF) / (LONG) displayWidth: 0;
+        *pcY = displayHeight ? ((y - 1) * 0xFFFF) / (LONG) displayHeight: 0;
     }
     else
     {
